backend/requirements.txt
# Core dependencies
flask>=2.3.3
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain dependencies with compatible versions
langchain>=0.1.0
langchain-huggingface>=0.1.2
langchain-core>=0.3.15,<0.4.0
langchain-community>=0.1.0
langgraph>=0.0.38

# Additional dependencies
pydantic>=2.5.2
typing-extensions>=4.9.0

# For LLM functionality
huggingface-hub>=0.23.0
sentence-transformers>=2.6.0
tokenizers>=0.19.1
transformers>=4.39.0

# For Web Searching
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.2 